Moves tour into a vue component + code cleanup
